feat(lobby): 20-game lobby with Parth cache optimizations, 58-asset manifest, and 3.5s warm launch
- Integrate all cache optimizations from Parth (ADR-007, ADR-008):
  * Stage-banded priority dispatch (PRELOADER -> COMMON -> SPLASH -> PRIMARY)
  * Session warm ledger keyed on exact, unnormalized URLs with retry on failure
  * Stream body draining to flush disk cache writes and bound bodies in flight
  * Priority hints (priority: 'low') for speculative background traffic
  * Connection pre-warming (<link rel=preconnect/dns-prefetch>) at boot
  * Slot-isolated URL prefix routing (/game/{id}_{session}/) in sandbox server
  * Foreground pressure cancellation (ADR-003) on game launch
- 58-asset critical startup manifest (28.87 MB) enabling green Play button at 3.5s warm
- Live descending ETA on card hover with visual priming indicators
- Operator controls: prefetch toggle, data budget selector, live telemetry HUD
- Robust fail-closed rollback handling for network drop simulation and timeouts
- Full test and benchmark harness passing 136/136 JS tests and end-to-end assertions

+++ tools/measure_har.py
@@ -11,9 +11,10 @@ from __future__ import annotations
 import argparse
 import json
 import math
+from collections.abc import Mapping, Sequence
 from datetime import datetime, timezone
 from pathlib import Path
-from typing import Any, Mapping, Sequence
+from typing import Any
 
 MEASURED = "MEASURED"
 UNKNOWN = "UNKNOWN"
@@ -109,7 +110,7 @@ def _wire_bytes(entry: Mapping[str, Any], classification: str) -> int | None:
 
     transfer_size = _nonnegative(response.get("_transferSize"))
     if transfer_size is not None:
-        return int(round(transfer_size))
+        return round(transfer_size)
     if classification == "hit":
         return 0
 
@@ -117,7 +118,7 @@ def _wire_bytes(entry: Mapping[str, Any], classification: str) -> int | None:
     headers_size = _nonnegative(response.get("headersSize"))
     body_size = _nonnegative(response.get("bodySize"))
     if headers_size is not None and body_size is not None:
-        return int(round(headers_size + body_size))
+        return round(headers_size + body_size)
     return None
